Commited idk, everything?

+++ backend-java/SpringCloudPXL/employee-service/src/main/java/be/pxl/services/repository/EmployeeRepository.java
@@ -3,6 +3,5 @@ package be.pxl.services.repository;
 import be.pxl.services.domain.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
-
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
